Switch back to 2 pages and put the data under the graph
